http-retry: move ReplayBody into a module
